# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Uptime Cairn API
  version: 1.0.0-draft
  summary: All-in-one open source uptime monitoring and reporting platform.
  description: |
    The complete public surface of Uptime Cairn. The dashboard is the first
    client of this API, never a privileged one — there are no internal
    endpoints the UI can reach that an API key cannot.

    ## Authentication

    Two schemes, either of which authenticates any non-public operation:

    - `ApiKeyAuth` — `Authorization: Bearer cairn_<key>`. Scoped, expiring,
      revocable. The scheme for automation, CI/CD, and Terraform.
    - `SessionCookie` — the browser session the dashboard uses after
      `POST /api/v1/auth/login`. Cookie-authenticated state-changing requests
      additionally require a CSRF token in `X-Cairn-CSRF-Token`.

    Operations tagged `Public` require no authentication: the status page read
    path, subscriber confirmation, and push-monitor ingest.

    ## Pagination

    Every list endpoint paginates by opaque cursor, keyed on `(updated_at, id)`,
    per ADR-004 — uniformly, with no exception for installs small enough that
    sending everything would "work". Offsets are unstable against rows that
    change while a user is paginating, and status changes constantly.

    Cursor responses carry no total count. A count for the current filter comes
    from `GET /api/v1/monitors/membership`, which a client polling for live-view
    freshness is already calling.

...

    ## Live updates
...
servers:
...
security:
  - ApiKeyAuth: []
  - SessionCookie: []
...
paths:
...
  /api/v1/imports/kuma:
    post:
      tags: [Import]
      operationId: importFromKuma
      summary: Import from one or more Uptime Kuma databases
      description: |
        Accepts one or several `kuma.db` files and reproduces monitors, tags,
        notification channels, and status pages. Supplying several files performs the
        multi-instance merge — the migration path for everyone currently sharding Kuma
        by hand across hosts.

        Import runs asynchronously. Poll the returned job for its status and report;
        nothing is silently dropped, and `dry_run` produces the same report without
        writing anything.
      x-cairn-phase: 1
      x-cairn-scopes: [imports:write]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/KumaImportRequest" }
            encoding:
              files:
                contentType: application/octet-stream
              options:
                contentType: application/json
      responses:
        "202":
          description: Import accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/imports/{importJobId}:
    parameters:
      - name: importJobId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Import]
      operationId: getImportJob
      summary: Retrieve an import job and its report
      description: |
        The report names every source entity and its outcome — imported, renamed,
        skipped, or failed, with the reason. This is the artefact a migrating user
        reads to confirm nothing was lost.
      x-cairn-phase: 1
      x-cairn-scopes: [imports:write]
      responses:
        "200":
          description: The import job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------------------
  # Settings
  # --------------------------------------------------------------------------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: cairn_<key>
      description: |
        A scoped API key: `Authorization: Bearer cairn_<key>`. Keys carry an explicit
        scope list, an optional expiry, and a last-used timestamp, and can be revoked
        at any time. Every operation names the scopes it requires in `x-cairn-scopes`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: cairn_session
      description: |
        The browser session established by `POST /api/v1/auth/login`. The cookie is
        `HttpOnly`, `SameSite=Lax`, and `Secure` when served over HTTPS.
        State-changing requests authenticated this way must also carry a matching
        `X-Cairn-CSRF-Token` header.
...
  responses:
...
    Unauthorized:
      description: No credentials were supplied, or they were not valid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PayloadTooLarge:
      description: The upload exceeded the configured limit.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationFailed:
      description: |
        The request was well-formed but semantically invalid. `errors` names each
        offending field by JSON pointer.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblem" }
...
  schemas:
...
    Problem:
      type: object
      description: An error, per RFC 9457.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            A stable URI identifying the error kind. Clients branch on this, never on
            `title` or `detail`, both of which are prose and may be translated.
          examples: ["https://uptimecairn.dev/errors/monitor-not-found"]
        title:
          type: string
          description: A short, human-readable summary of the error kind.
        status:
          type: integer
          description: The HTTP status code, repeated for clients that only read the body.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: The request path that produced the error.
        request_id:
          type: string
          description: Correlates this response with the server log entry.
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              description: One entry per invalid field.
              items:
                type: object
                required: [pointer, code, message]
                properties:
                  pointer:
                    type: string
                    description: RFC 6901 JSON pointer to the offending field.
                    examples: ["/config/interval_seconds"]
                  code:
                    type: string
                    description: Stable machine-readable reason.
                    examples: ["below_minimum"]
                  message:
                    type: string
                    description: Human-readable explanation.
...
    KumaImportRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          minItems: 1
          description: |
            One or more Uptime Kuma SQLite databases. Several files performs the
            multi-instance merge.
          items: { type: string, format: binary }
        options:
          type: object
          properties:
            dry_run:
              type: boolean
              default: false
              description: Produce the full report without writing anything.
            conflict_strategy:
              type: string
              enum: [skip, rename, replace]
              default: rename
              description: |
                What to do when an imported entity collides with an existing one by name.
                `rename` suffixes the incoming entity, which is the only option that
                cannot lose data.
            name_prefix:
              type: [string, "null"]
              description: |
                Prefix applied to imported monitor names. The practical way to keep several
                merged Kuma instances distinguishable afterwards.
            import_monitors: { type: boolean, default: true }
            import_tags: { type: boolean, default: true }
            import_notifications: { type: boolean, default: true }
            import_status_pages: { type: boolean, default: true }
            import_history:
              type: boolean
              default: false
              description: Import historical heartbeats as well as configuration. Slower, and larger.
            enable_after_import:
              type: boolean
              default: false
              description: |
                Leave imported monitors paused by default, so a migrating user reviews
                them before five thousand checks start firing at once.
    ImportJob:
      type: object
      required: [id, state, created_at]
      properties:
        id: { type: string, format: uuid }
        state: { type: string, enum: [queued, running, succeeded, partial, failed] }
        dry_run: { type: boolean }
        sources:
          type: array
          description: One entry per uploaded database.
          items:
            type: object
            properties:
              filename: { type: string }
              kuma_version: { type: [string, "null"] }
              detected_entities:
                type: object
                additionalProperties: { type: integer }
        summary:
          type: object
          description: Counts by entity type and outcome.
          additionalProperties:
            type: object
            properties:
              imported: { type: integer }
              renamed: { type: integer }
              skipped: { type: integer }
              failed: { type: integer }
        entries:
          type: array
          description: |
            Per-entity outcomes. Every source entity appears exactly once — the guarantee
            that nothing was silently dropped.
          items:
            type: object
            required: [source_file, entity_type, source_name, result]
            properties:
              source_file: { type: string }
              entity_type: { type: string, enum: [monitor, tag, notification, status_page, heartbeat_range] }
              source_id: { type: [string, "null"] }
              source_name: { type: string }
              result: { type: string, enum: [imported, renamed, skipped, failed, unsupported] }
              target_id: { type: [string, "null"], format: uuid }
              detail:
                type: [string, "null"]
                description: |
                  Why, in plain language — an unsupported monitor type, a notification
                  provider with no Cairn equivalent, a name collision and what was done about it.
        error: { type: [string, "null"] }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }

    # ------------------------------------------------------------------------
    # Settings and users
    # ------------------------------------------------------------------------
...